dev: use `protobuf` instead of `exprotobuf`
The exprotobuf library appears to be abandoned. In Elixir 1.18,
warning logs are being produced at runtime. So, switch to the most
popular protobuf library on hex.pm and make the needed adjustments.

Also the progression service needs this so it does not conflict with
blake_protobufs `2.0.1`.

Prep for a 2.0.0 release. Details of the breaking change are in the
CHANGELOG.md.